+++ app/api/staff.ts
@@ -1,4 +1,4 @@
-import { useMutation, useQuery } from "react-query";
+import { useMutation, useQuery, useQueryClient } from "react-query";
 import { apiCall } from "@/utils/apiRequest";
 import endpoints from "@/utils/endpoints";
 import { useAppDispatch } from "@/hooks";
@@ -6,7 +6,7 @@ import { setMessage, setShowToast } from "@/store/toastSlice";
 import { DynamicObject } from "@/components/types";
 
 export const useGetAllStaff = (slug: string) => {
-  return useQuery("staff", async () => {
+  return useQuery("getAllStaff", async () => {
     try {
       const response = await apiCall(
         "GET",
@@ -36,36 +36,47 @@ export const useCreateStaff = () => {
 };
 
 export const useDeleteStaff = () => {
+  const queryClient = useQueryClient();
   const dispatch = useAppDispatch();
-  return useMutation<void, Error, any>(async (staff_id: number) => {
-    const response = await apiCall(
-      "DELETE",
-      `${endpoints.deleteStaff}${staff_id}`,
-      {},
-      {}
-    );
-    dispatch(setMessage(response.message));
-    dispatch(setShowToast(true));
-    return response;
-  });
-};
-
-export const useEditStaff = () => {
-  const dispatch = useAppDispatch();
-  return useMutation<void, Error, DynamicObject>(
-    async ({ f_name, phone, role, staff_id }) => {
+  return useMutation<void, Error, any>(
+    async (staff_id: number) => {
       const response = await apiCall(
-        "PUT",
-        `${endpoints.editStaff}${staff_id}`,
-        { f_name, phone, role },
+        "DELETE",
+        `${endpoints.deleteStaff}${staff_id}`,
+        {},
         {}
       );
       dispatch(setMessage(response.message));
       dispatch(setShowToast(true));
-      setTimeout(() => {
-        dispatch(setMessage(""));
-      }, 3000);
       return response;
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries("getAllStaff");
+      },
+    }
+  );
+};
+
+export const useEditStaff = () => {
+  const queryClient = useQueryClient();
+  const dispatch = useAppDispatch();
+  return useMutation<void, Error, any>(
+    async ({ id, phone, role }) => {
+      const response = await apiCall(
+        "PUT",
+        `${endpoints.editStaff}${id}`,
+        { phone, role },
+        {}
+      );
+      dispatch(setMessage(response.message));
+      dispatch(setShowToast(true));
+      return response;
+    },
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries("getAllStaff");
+      },
     }
   );
 };
